Moved all files to new namespace com.sirenofshame.soscmd
