+++ src/main/java/com/sirenofshame/soscmd/HardwareType.java
@@ -1,4 +1,4 @@
-package com.hanik.usb.siren;
+package com.sirenofshame.soscmd;
 
 public enum HardwareType {
         /// <summary>
