+++ src/main/java/com/sirenofshame/soscmd/ManualControlData.java
@@ -1,4 +1,4 @@
-package com.hanik.usb.siren;
+package com.sirenofshame.soscmd;
 
 public class ManualControlData {
     public byte Led0;
