+++ src/main/resources/99-sirenusb.rules
@@ -1,1 +1,1 @@
-SUBSYSTEM=="usb",ATTR{idVendor}=="16d0",ATTR{idProduct}=="0646",MODE="0660",GROUP="fhanik"
+SUBSYSTEM=="usb",ATTR{idVendor}=="16d0",ATTR{idProduct}=="0646",MODE="0660",GROUP="sirenofshame"
